blm bisa edit proker

+++ Tugas-Besar/fosma/listsejarah.php
@@ -20,7 +20,7 @@ $data->title
 $data->deskripsi";
 echo "
 <br>
-<a class='btn btn-danger' href='listsejarah.php?delete=$data->title'>Delete</a>
+<a class='btn btn-danger' href='listsejarah.php?delete=$data->title' onclick='return confirm(\"Apakah anda yakin ingin menghapus?\")'>Delete</a>
 <a class='btn btn-info' href='editsejarah.php?title=$data->title'>Edit</a>
 ";
         };
